requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Training dependencies
pytorch-lightning==2.1.3
tensorboard==2.15.1
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0